Enabled automatic alarm Added support for 	- populating EVENT row 	- removing events from Event row if a threshold is reached 	- removing RDR events (only those are marked for deletion - via the RPT populate_rpt call) 	- removing RTP events

+++ snmp-subagent/trunk/include/saHpiTable.h
@@ -141,7 +141,7 @@ extern          "C" {
 
   int populate_rpt( void );
 
-  int purge_rpt( void );
+  unsigned long purge_rpt( void );
   /* 
    * The various SET operations. 
    */
